RUMM-712 Split `Main.storyboard` into separate storyboards,
one per each test scenario (Logging, Tracing, RUM).
