+++ Datadog/Example/Scenarios/Logging/LoggingScenario.storyboard
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16097.2" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="Jlq-WO-Qzl">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Send Logs Fixture View Controller-->
+        <scene sceneID="TXw-Wg-Kee">
+            <objects>
+                <viewController id="Jlq-WO-Qzl" customClass="SendLogsFixtureViewController" customModule="Example" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="cdr-c3-9K3">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text=" Sending logs... " textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="4yI-pe-L5M">
+                                <rect key="frame" x="146" y="437.5" width="122" height="21"/>
+                                <color key="backgroundColor" red="0.38823529410000002" green="0.17254901959999999" blue="0.65098039220000004" alpha="1" colorSpace="calibratedRGB"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="4yI-pe-L5M" firstAttribute="centerX" secondItem="cdr-c3-9K3" secondAttribute="centerX" id="Tqb-5A-Dxp"/>
+                            <constraint firstItem="4yI-pe-L5M" firstAttribute="centerY" secondItem="cdr-c3-9K3" secondAttribute="centerY" id="rot-Hg-hZg"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="gNT-5l-zGn"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="gZZ-Us-17W"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="PLr-Gj-bpm" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-1386" y="1233"/>
+        </scene>
+    </scenes>
+</document>
