+++ Datadog/Example/Scenarios/Tracing/TracingScenario.storyboard
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16097.2" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="KCH-2G-uv6">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Send Traces Fixture View Controller-->
+        <scene sceneID="opo-fL-CWd">
+            <objects>
+                <viewController id="KCH-2G-uv6" customClass="SendTracesFixtureViewController" customModule="Example" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="NrI-Vk-ZoQ">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text=" Sending traces... " textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ecs-Xp-TM2">
+                                <rect key="frame" x="138" y="437.5" width="138" height="21"/>
+                                <color key="backgroundColor" red="0.38823529410000002" green="0.17254901959999999" blue="0.65098039220000004" alpha="1" colorSpace="calibratedRGB"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="ecs-Xp-TM2" firstAttribute="centerX" secondItem="NrI-Vk-ZoQ" secondAttribute="centerX" id="SWO-zZ-aoq"/>
+                            <constraint firstItem="ecs-Xp-TM2" firstAttribute="centerY" secondItem="NrI-Vk-ZoQ" secondAttribute="centerY" id="dG3-dd-OmP"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="dLW-Q9-mWI"/>
+                    </view>
+                    <navigationItem key="navigationItem" id="36v-HB-HqG"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="VmN-JW-kgu" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-542" y="1233"/>
+        </scene>
+    </scenes>
+</document>
